[refactor] #13 Calendar 관련 파일 패키징 작업 완료
